WIP bitwarden integration
Fixed continues call to AddConnection instead of EditConnection for
bitwarden storage.

Set private key and public key in item if password is not provided.
Introduced plink.exe and sshpass to login into ssh by providing inline
the password to it.

TODO: when trying to log into an ssh session with keyfile use the one
from the item.